requirements_dev.txt
gunicorn==20.0.0
jupyter==1.0.0
matplotlib==3.3.2
nbgrader==0.6.1
notebook==6.1.4
numpy==1.19.4
pandas==1.1.4
scikit-learn==0.23.2
category_encoders==2.2.2
flask==1.1.2
peewee==3.13.3
psycopg2-binary==2.8.6
requests==2.24.0
seaborn==0.11.0
alembic==1.4.2
appnope==0.1.0
argon2-cffi==20.1.0
attrs==20.1.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fuzzywuzzy==0.18.0
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
kiwisolver==1.2.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
nbgrader==0.6.1
notebook==6.1.3
numpy==1.19.1
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pmdarima==1.7.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.19
statsmodels==0.11.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.49.0
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
nltk==3.5
numpy==1.19.2
scipy==1.5.2
seaborn==0.11.0
spacy==2.3.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.0/en_core_web_md-2.3.0.tar.gz#egg=en_core_web_md


